+++ docs/CHANGELOG.md
@@ -7,7 +7,7 @@ The format is (read: strives to be) based on [Keep a Changelog](https://keepacha
 ## [1.2.18] - 2026-04-06
 ### Changed:
 - Add yes flag to the remove command, to avoid a headless prompt hang.
-- Add vault functions wrappers to DworshakSecret class, to pass the db_path attribute.
+- Add vault functions wrappers to DworshakSecret class, to pass the db_path attribute. Leave out import function, which requires a JSON path.
 
 ### Internal:
 - I am so close to tossing out non-Typer CLI. It is not worth the maintenance.