+++ docs/CHANGELOG.md
@@ -7,10 +7,11 @@ The format is (read: strives to be) based on [Keep a Changelog](https://keepacha
 ## [1.2.18] - 2026-04-06
 ### Changed:
 - Add yes flag to the remove command, to avoid a headless prompt hang.
-- Add vault functions wrappers to DworshakSecret class, to pass the db_path attribute. Leave out import function, which requires a JSON path.
+- Add vault functions wrappers to DworshakSecret class; leverage `**kwargs` to passthrough vars without increasing the maintenance burden For the import_records() function wrapper, explicitly provide the json_path keyword, in addition to the `**kwargs` passthrough, without which the function is useless.
 
 ### Internal:
 - I am so close to tossing out non-Typer CLI. It is not worth the maintenance.
+- Move exposure of ecosystem libs from dworshak to dworshak-prompt. dworshak is a cli, that is all. Users should not expect to have dworshak in a project venv, only in their system via pipx.
 
 ---